Plataforma B3/src/requirements.txt
# FastAPI e dependências
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3

# Azure
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-ai-formrecognizer==3.3.0
azure-ai-textanalytics==5.3.0
azure-ai-translation==1.0.0
azure-ai-vision==0.11.0
azure-ai-language==1.0.0
azure-ai-metricsadvisor==1.0.0
azure-ai-personalizer==1.0.0
azure-ai-anomalydetector==3.0.0
azure-ai-contentsafety==1.0.0
azure-ai-documentintelligence==1.0.0
azure-ai-generative==1.0.0
azure-ai-metricsadvisor==1.0.0
azure-ai-personalizer==1.0.0
azure-ai-translation==1.0.0
azure-ai-vision==0.11.0

# OpenAI
openai==1.3.0
tiktoken==0.5.1

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0 